requirements.txt
pkg-resources==0.0.0
pygogo==0.12.0
python-dotenv==0.8.2
selenium==2.53.6
